Add nnz parameter to sparse.random
Currently, `sparse.random` uses the `density` argument to determine
the number of non-zero elements in the generated array. If the exact
number of non-zero elements is required, one might try to compute
density as `nnz / np.prod(shape)`, which might give a wrong result
for large values due to a floating-point error.

This PR adds a new, optional `nnz` argument to `sparse.random` that,
if specified, overrides `density` and allows a caller to create
a random array with the exact number of non-zero elements.